backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
sqlalchemy==2.0.40
pydantic-settings==2.8.1
python-dotenv==1.1.0
pytest==8.3.5
httpx==0.28.1
